CreatePoapContext / POAP creation flow refactor
